# docs/swagger.yaml
definitions:
  model.CreateDraftResponse:
    properties:
      draft_id:
        example: 1
        type: integer
    type: object
  model.ResponseResult:
    properties:
      code:
        example: 0
        type: integer
      message:
        example: ok
        type: string
      result: {}
    type: object
info:
  contact: {}
  description: 玩起来！！！
  title: go-toy
  version: "1.0"
paths:
  /api/ping:
    get:
      produces:
      - application/json
      responses:
        "0":
          description: ok
          schema:
            type: string
      summary: 测试Ping
  /changePassword:
    post:
      consumes:
      - application/json
      parameters:
      - description: 旧密码
        in: body
        name: oldPassword
        required: true
        schema:
          type: string
      - description: 新密码
        in: body
        name: newPassword
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 修改密码
      tags:
      - user
  /createDraft:
    get:
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseResult'
            - properties:
                result:
                  $ref: '#/definitions/model.CreateDraftResponse'
              type: object
      summary: 创建草稿
      tags:
      - 草稿
  /deleteDraft:
    post:
      consumes:
      - application/json
      parameters:
      - description: 草稿id
        in: body
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 删除草稿
      tags:
      - 草稿
  /getDraft:
    get:
      parameters:
      - description: 草稿id
        in: body
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 获取草稿
      tags:
      - 草稿
  /getDraftList:
    get:
      parameters:
      - description: 页数
        in: query
        name: page
        type: integer
      - description: 页码
        in: query
        name: page_size
        type: integer
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 获取草稿箱
      tags:
      - 草稿
  /getMenuList:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 获取角色菜单
      tags:
      - menu
  /getUserInfo:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 获取用户信息
      tags:
      - user
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 登录
      tags:
      - user
  /publishDraft:
    post:
      parameters:
      - description: 草稿id
        in: body
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 发布
      tags:
      - 草稿
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 注册
      tags:
      - user
  /saveDraft:
    post:
      consumes:
      - application/json
      parameters:
      - description: 草稿id
        in: body
        name: id
        required: true
        schema:
          type: integer
      - description: 标题
        in: body
        name: title
        schema:
          type: string
      - description: 内容
        in: body
        name: content
        schema:
          type: string
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/model.ResponseResult'
      summary: 保存草稿
      tags:
      - 草稿
  /upload/file:
    post:
      consumes:
      - application/form-data
      parameters:
      - description: 类型
        in: body
        name: type
        required: true
        schema:
          type: string
      - description: 文件
        in: body
        name: file
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses: {}
      summary: 上传文件
      tags:
      - 基建
  /visit:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 访问网页, 埋点上报
      tags:
      - 基建
swagger: "2.0"